requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pillow==10.4.0
numpy==1.26.4
scikit-image==0.24.0
h5py==3.11.0
SQLAlchemy==2.0.23
imagehash==4.3.1
tensorflow==2.16.1
keras==3.3.3
h5py==3.11.0
pillow==10.4.0
scikit-image==0.24.0
imagehash==4.3.1


